Another set of API breaking changes to introduce array of instances.
1.
Changes to Actions/Action : now both concepts are captured by a Statement.
Basically renaming Actions to Statement should work for pure GAL users.
This means for loops are possible in composite declarations...

Class Call (was a self call for GAL) is now SelfCall, fused into the concept of Composite.

2.
Fused various kinds of instances into InstanceDeclaration, which has a type and a name.
New ArrayInstanceDeclaration has type, name and size for arrays of instances.
InstanceDecl : ArrayInstanceDeclaration | InstanceDeclaration
And instances of a CompositeTD are now VarDecl.
This change to instances loses the type of the instance so code such as 
if (obj instanceof GalInstance) { gali.getType() .... }
become
if (obj instanceof InstanceDeclaration && ((InstanceDeclaration)obj).getType() instanceof GALTypeDeclaration ) { obj.getType() .... }

InstanceCall instance parameter is now a Reference allowing to refer to cell of an ArrayInstance.