Refined Read=>Feed : we need a *real* feeder that does *not* eat
In case of a reader/feeder, e.g. t consumes -1 and puts +2 in place p,
we mistakenly counted t as both a reader (ok, despite the overall effect
at +1, we do read) and feeder (NOK, we cannot be the one that "unlocks"
place p).

This means we would build constraints of the form :
`t16 > 0 => t0 >0 OR t16 > 0`

which don't actually constrain anything (they are not wrong and they
don't hurt at least).

Also refactor to use `map.computeIfAbsent` in a couple places.