Patch Parikh Walk + Add new Heuristic MIN/MAX
* Now editing Parikh vector count BEFORE discarding unavailable, might
impact in some cases where last fired transition is still enabled (or
one of its representatives) it would not have been discarded, which does
not help realize our Parikh vector
* New heuristic : try to take the last or first transition available.
Due to the way the enabled transitions are managed, this leads to
chaining behavior when available (MAX case), which might actually be a
good idea. Finds some traces we couldn't reproduce before for B&V model.