Two major changes, one API breaking. Performance patch.
O( factorial height of syntax tree ) more or less to O (Nodes in tree)
1. Simplify is now a non static method, so that it can benefit from a
cache. The problem was that we paid N traversals of subtrees in
simplify, with N the height of the syntactic tree... The cache avoids
this pitfall.

2. We no longer "push" negations down within simplify, that can (and is)
be done later. Current procedure was terribly inefficient for high
negations in the syntax tree. Normal algorithm uses a boolean that
changes value as Not nodes are traversed.