FIX : patch Trap constraints : was erroneously forbidding some
this caused less trap constraints to be found, but traps which were
found were still correct in all cases. So we increase our decision power
now.

solution vector indexes are no longer consistent with current net as
soon as we start discarding parts of it. We already dropped all marked
places in the solution anyway.