A lightweight AST for heavy lifting
Main API is getOp + two internal nodes and heaps of leaves that mostly
let you "getValue" polymorphically. Some refactoring to fuse some leaves
would be feasible.

EMF *is* a nice framework, and I *still* do like it.
But sometimes you need to move on...or go back to your basics as you
will.