Adopt Bloom filter from @lovasoa project (Thanks !)
See source project : https://github.com/lovasoa/bloomfilter

It is used for a prototype "probabilistic verification" approach, per
the terminology of
"A New Scheme for Memory-Efficient Probabilistic Verification (1996)",
Stern & Dill

or earlier work by Holzmann ('87) working on Spin that is cited in that
paper, and that he later worked on again e.g. "An Analysis of Bitstate
Hashing (1995)"

This is not actual probabilistic model-checking with stochastic models
or probabilities.