Use DoneProperties instead of Map<String,Bool>
This object may now serve as an observer/notifu=ication strategy to the
computation progress, with some code triggered when we add a new result.
Previously code would Sysout the result and separately add to the Map.

Allows to better control what we do with the results. Opens the road to
more reuse of the code base outside of producing MCC compliant traces.

Hopefully this huge commit is not breaking anything, but it sure is
intrusive. Otoh it simplifies the code (a single place now with the
keyword "FORMULA"...) and may avoid some bugs (doneProps and sysout not
in accord like we detected in a recent patch).