Performance : limit complexity when number of formulas grows
NEW : only test one out of 100 states met when over 10k expressions to
check.

This could be an issue for QuasiLive or OneSafe for instance, testing
each property at every step means we never reach steps that are deeper
than a few steps in.

Also use a parallel evaluation of expressions in each state reached to
get better throughput.