Be more careful with guarded transitions and skeleton
A property CTL/LTL that has some combination of

t1 and t2 enabled

where t1 and/or t2 is guarded,

and the preset of t1 and t2 overlap

is suspicious, because we might e.g. think they are always both enabled
when one is enabled in the skeleton, when this might be untrue due to
guards in the HLPN model.

Patches (at least) two bugs in MCC 2021 LTLF category :
NeoElection-COL-5-LTLFireability-09
BridgeAndVehicles-COL-V50P20N50-LTLFireability-11