adding support for check-sat assuming to JSMTlib
This support is not implemented even in the upstream
https://github.com/smtlib/jSMTLIB of @davidcok, where part of the
command exists in the code base, but with wrong API (should be a List of
IExpr ? symbols or their negation only) and no actual Solver support.

TODO: maybe propose this commit upstream as a clean PR, but the sources have diverged since my fork.